pyproject.toml
[project]
name = "bg-database"
version = "0.1.0"
description = ""
authors = [
    { name = "ryouhama" }
]
dependencies = [
    "pydantic>=2.4.2",
    "pytest-mock>=3.12.0",
    "supabase>=2.0.3",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.3",
]

[tool.rye.workspace]
members = ["app/*", "test/*"]

[tool.hatch.metadata]
allow-direct-references = true
